feat: Update the encrypted note log format
