+++ noir-projects/noir-contracts/contracts/token_blacklist_contract/src/types/token_note.nr
@@ -51,7 +51,7 @@ impl NoteInterface<TOKEN_NOTE_LEN> for TokenNote {
       // We only bother inserting the note if non-empty to save funds on gas.
       // TODO: (#5901) This will be changed a lot, as it should use the updated encrypted log format
       if !(self.amount == U128::from_integer(0)) {
-          context.emit_note_encrypted_log(
+          context.encrypt_and_emit_note(
               (*context).this_address(),
               slot,
               Self::get_note_type_id(),
