+++ yarn-project/circuit-types/src/logs/index.ts
@@ -10,6 +10,6 @@ export * from './l1_note_payload/index.js';
 export * from './tx_l2_logs.js';
 export * from './unencrypted_l2_log.js';
 export * from './extended_unencrypted_l2_log.js';
-export * from './encrypted_log_header.js';
-export * from './encrypted_log_incoming_body.js';
-export * from './encrypted_log_outgoing_body.js';
+export * from './l1_note_payload/encrypted_log_header.js';
+export * from './l1_note_payload/encrypted_log_incoming_body.js';
+export * from './l1_note_payload/encrypted_log_outgoing_body.js';
