+++ yarn-project/circuit-types/src/logs/l1_note_payload/encrypted_log_header.test.ts
